test: code refactoring
