--- src/components/AutosuggestDeprecated/ItemTag/ItemTag.scss
@@ -1,31 +1,0 @@
-.ItemTag {
-    display: inline-flex;
-    justify-content: space-between;
-    align-items: baseline;
-    font-size: var(--font-size-small);
-    padding: var(--spacing-normal) var(--spacing-2x);
-    background-color: var(--color-neutral-25);
-    border-radius: var(--border-radius);
-    margin: var(--spacing-normal);
-    min-width: 100px;
-    max-width: 150px;
-    flex: 1;
-    cursor: default;
-
-    &__label {
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        padding: 0 var(--spacing-2x) 0 0;
-    }
-
-    &__button {
-        border: 0;
-        cursor: pointer;
-        background-color: var(--transparent);
-        padding: 0 0 0 0;
-        font-size: var(--font-size-large);
-        line-height: var(--font-size-small);
-        font-weight: var(--font-weight-bold);
-    }
-}
