--- src/components/AutosuggestDeprecated/index.js
@@ -1,1 +1,0 @@
-export { ItemTag } from './ItemTag';
