+++ src/components/SelectComponents/ComboboxMulti/__tests__/ComboboxMulti.spec.tsx
@@ -3,10 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { ComboboxMulti } from '../ComboboxMulti';
-import {
-    SUGGESTIONS,
-    SUGGESTION_TO_STRING,
-} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
+import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../../constants/suggestions';
 
 describe('ComboboxMulti', () => {
     const suggestionToString = SUGGESTION_TO_STRING;
