+++ src/components/SelectComponents/Select/__tests__/Select.spec.tsx
@@ -3,10 +3,7 @@ import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { Select } from '../Select';
-import {
-    SUGGESTIONS,
-    SUGGESTION_TO_STRING,
-} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
+import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../../constants/suggestions';
 
 // TODO: these tests use the wrong role to find items in the list.
 // This should be remedied in or following https://textkernel.atlassian.net/browse/ONEUI-364
