+++ src/components/SelectComponents/SelectBase/__tests__/SelectBase.spec.tsx
@@ -4,10 +4,7 @@ import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { SelectBase } from '../SelectBase';
 import { SuggestionsList } from '../../SuggestionsList';
-import {
-    SUGGESTIONS,
-    SUGGESTION_TO_STRING,
-} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
+import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../../constants/suggestions';
 
 describe('SelectBase', () => {
     const inputRef = React.createRef();
