+++ src/index.ts
@@ -63,7 +63,6 @@ export * from './components/WeightedResultBar';
 export * from './components/NumericStepper';
 // Organisms
 export * from './components/Dialogs';
-export * from './components/AutosuggestDeprecated';
 export * from './components/SelectComponents';
 export * from './components/LocationAutocomplete';
 export * from './components/LocationSelector';
